+++ pydocumentdb/errors.py
@@ -1,4 +1,23 @@
-﻿# Copyright (c) Microsoft Corporation.  All rights reserved.
+﻿#The MIT License (MIT)
+#Copyright (c) 2014 Microsoft Corporation
+
+#Permission is hereby granted, free of charge, to any person obtaining a copy
+#of this software and associated documentation files (the "Software"), to deal
+#in the Software without restriction, including without limitation the rights
+#to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+#copies of the Software, and to permit persons to whom the Software is
+#furnished to do so, subject to the following conditions:
+
+#The above copyright notice and this permission notice shall be included in all
+#copies or substantial portions of the Software.
+
+#THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+#IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+#FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+#AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+#LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+#OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+#SOFTWARE.
 
 """PyDocumentDB Exceptions.
 """
@@ -25,6 +44,7 @@ class HTTPFailure(DocumentDBError):
         self.status_code = status_code
         self.headers = headers
         self.sub_status = None
+        self._http_error_message = message
         if http_constants.HttpHeaders.SubStatus in self.headers:
             self.sub_status = int(self.headers[http_constants.HttpHeaders.SubStatus])
             DocumentDBError.__init__(self,
@@ -32,7 +52,6 @@ class HTTPFailure(DocumentDBError):
         else:
             DocumentDBError.__init__(self,
                                  'Status code: %d\n%s' % (self.status_code, message))
-
 
 class JSONParseFailure(DocumentDBError):
     """Raised when fails to parse JSON message.